+++ src/test/java/de/wbstraining/lotto/populatedb/PopulateSpielUndGewinnklasse.java
@@ -5,7 +5,6 @@
  */
 package de.wbstraining.lotto.populatedb;
 
-import java.math.BigInteger;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.Month;
@@ -24,14 +23,14 @@ import de.wbstraining.lotto.persistence.model.Spiel;
  */
 @Stateless
 public class PopulateSpielUndGewinnklasse
-	implements PopulateSpielUndGewinnklasseLocal {
+		implements PopulateSpielUndGewinnklasseLocal {
 	@EJB
 	private SpielFacadeLocal spielFacade;
 	@EJB
 	private GewinnklasseFacadeLocal gewinnklasseFacade;
 
 	private void populateGewinnklassen(LocalDate gueltigAb, LocalDate gueltigBis,
-		Spiel spiel, String[] gewinnKlassen) {
+			Spiel spiel, String[] gewinnKlassen) {
 
 		Gewinnklasse gewinnklasse;
 		String[] items;
@@ -50,7 +49,7 @@ public class PopulateSpielUndGewinnklasse
 			gewinnklasse.setBeschreibung(items[0]);
 			gewinnklasse.setGewinnklassenr(Integer.parseInt(items[1]));
 			gewinnklasse.setBezeichnunglatein(items[2]);
-			gewinnklasse.setBetrag(BigInteger.valueOf(Long.parseLong(items[3])));
+			gewinnklasse.setBetrag(Long.parseLong(items[3]));
 			gewinnklasse.setIsabsolut(Boolean.valueOf(items[4]));
 			gewinnklasse.setGueltigab(gueltigAb);
 			gewinnklasse.setGueltigbis(gueltigBis);
@@ -67,38 +66,38 @@ public class PopulateSpielUndGewinnklasse
 		// betrag
 		// isAbsolut
 		String[] gewinnklassen6Aus49 = { "6 Richtige + SZ,1,I,1280,false",
-			"6 Richtige,2,II,1000,false", "5 Richtige + SZ,3,III,500,false",
-			"5 Richtige,4,IV,1500,false", "4 Richtige + SZ,5,V,500,false",
-			"4 Richtige,6,VI,1000,false", "3 Richtige + SZ,7,VII,1000,false",
-			"3 Richtige,8,VIII,4500,false", "2 Richtige + SZ,9,IX,500,true" };
+				"6 Richtige,2,II,1000,false", "5 Richtige + SZ,3,III,500,false",
+				"5 Richtige,4,IV,1500,false", "4 Richtige + SZ,5,V,500,false",
+				"4 Richtige,6,VI,1000,false", "3 Richtige + SZ,7,VII,1000,false",
+				"3 Richtige,8,VIII,4500,false", "2 Richtige + SZ,9,IX,500,true" };
 		// beschreibung
 		// gewinnklasseNr
 		// bezeichnungLatein
 		// betrag
 		// isAbsolut
 		String[] gewinnklassenSpiel77 = { "richtige Gewinnzahl,1,I,0,true",
-			"6 richtige Endziffern,2,II,7777700,true",
-			"5 richtige Endziffern,3,III,777700,true",
-			"4 richtige Endziffern,4,IV,77700,true",
-			"3 richtige Endziffern,5,V,7700,true",
-			"2 richtige Endziffern,6,VI,1700,true",
-			"1 richtige Endziffer,7,VII,500,true" };
+				"6 richtige Endziffern,2,II,7777700,true",
+				"5 richtige Endziffern,3,III,777700,true",
+				"4 richtige Endziffern,4,IV,77700,true",
+				"3 richtige Endziffern,5,V,7700,true",
+				"2 richtige Endziffern,6,VI,1700,true",
+				"1 richtige Endziffer,7,VII,500,true" };
 		// beschreibung
 		// gewinnklasseNr
 		// bezeichnungLatein
 		// betrag
 		// isAbsolut
 		String[] gewinnklassenSuper6 = { "6 richtige Endziffern,1,I,10000000,true",
-			"5 richtige Endziffern,2,II,666600,true",
-			"4 richtige Endziffern,3,III,66600,true",
-			"3 richtige Endziffern,4,IV,6600,true",
-			"2 richtige Endziffern,5,V,600,true",
-			"1 richtige Endziffer,6,VI,250,true" };
+				"5 richtige Endziffern,2,II,666600,true",
+				"4 richtige Endziffern,3,III,66600,true",
+				"3 richtige Endziffern,4,IV,6600,true",
+				"2 richtige Endziffern,5,V,600,true",
+				"1 richtige Endziffer,6,VI,250,true" };
 
 		LocalDateTime date = LocalDateTime.now();
 		LocalDate gueltigAb = LocalDate.of(2015, Month.JANUARY, 1);
 		LocalDate gueltigBis = gueltigAb.plusYears(10)
-			.minusDays(1);
+				.minusDays(1);
 
 		Spiel sechsAus49 = new Spiel();
 		Spiel super6 = new Spiel();
@@ -127,7 +126,7 @@ public class PopulateSpielUndGewinnklasse
 		spielFacade.create(spiel77);
 
 		populateGewinnklassen(gueltigAb, gueltigBis, sechsAus49,
-			gewinnklassen6Aus49);
+				gewinnklassen6Aus49);
 		populateGewinnklassen(gueltigAb, gueltigBis, spiel77, gewinnklassenSpiel77);
 		populateGewinnklassen(gueltigAb, gueltigBis, super6, gewinnklassenSuper6);
 	}
